Allow !isbanned @me to target self
Proposed change to support upcoming 2.0.0 of CTBans plugin. The current version does not support sm_isbanned @me but the 2.0.0 version will support this command as long as the convar sm_ctban_isbanned_selftarget is 1.  if sm_ctba_isbanned_selftarget is 0 the client would receive a message on how they do not have access to the command.